preprocessor/Hamer/hamer_requirements.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
controlnet-aux==0.0.9
cycler==0.12.1
Cython==0.29.37
datasets==2.20.0
deepspeed==0.14.4
# Editable install with no version control (detectron2==0.6)
#-e /data01/zhanghy/test_human_dev/detectron2
#diffusers==0.29.0
dill==0.3.8
docstring_parser==0.16
einops==0.8.0
embreex==2.17.7.post4
exceptiongroup==1.2.1
filelock==3.15.1
fonttools==4.53.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.5.0
fvcore==0.1.5.post20221221
grpcio==1.64.1
hjson==3.1.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
iniconfig==2.0.0
iopath==0.1.9
Jinja2==3.1.4
json-tricks==3.17.3
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.3.post0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mmcv==1.5.0
mmdet==3.3.0
mmengine==0.10.4
# Editable install with no version control (mmpose==0.24.0)
-e /nvfile-heatstorage/zangxh/intern/wangyx/codes_zhy_20240604/src/asdff/asdff/Hamer/third-party/ViTPose
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==2.3.0
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peft==0.11.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.7.4
pydantic_core==2.18.4
pyembree==0.1.12
pyglet==2.0.15
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pyrender==0.1.45
pytest==8.2.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.1
pytz==2024.1
PyYAML==6.0.1
rectpack==0.2.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
Rtree==1.2.0
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
shtab==1.7.1
six==1.16.0
smplx==0.1.28
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2024.5.22
timm==0.6.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.4.0.post0
torchvision==0.16.0
tqdm==4.66.4
transformers==4.41.2
#trimesh==4.0.5
triton==2.1.0
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
ultralytics==8.2.35
ultralytics-thop==2.0.0
urllib3==2.2.1
webdataset==0.2.86
Werkzeug==3.0.3
xtcocotools==1.14.3
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
